# data-layer/openapi.json
{
  "components": {
    "schemas": {
...
      "AskRequest": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Row cap for the answer.",
            "title": "Limit"
          },
          "question": {
            "description": "A question about the fleet, in English.",
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "BatchQueryRequest": {
        "properties": {
          "id": {
            "description": "Your name for this query; it keys the results.",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "limit": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Row cap for this query.",
            "title": "Limit"
          },
          "sql": {
            "description": "A SELECT, possibly with {{variable}} placeholders.",
            "minLength": 1,
            "title": "Sql",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sql"
        ],
        "title": "BatchQueryRequest",
        "type": "object"
      },
      "BatchRequest": {
        "properties": {
          "params": {
            "additionalProperties": true,
            "description": "Values for the {{placeholders}}, substituted as SQL literals.",
            "title": "Params",
            "type": "object"
          },
          "queries": {
            "items": {
              "$ref": "#/components/schemas/BatchQueryRequest"
            },
            "maxItems": 24,
            "minItems": 1,
            "title": "Queries",
            "type": "array"
          }
        },
        "required": [
          "queries"
        ],
        "title": "BatchRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "properties": {
          "params": {
            "additionalProperties": true,
            "description": "Values for the dashboard's variables; missing ones fall back to the default.",
            "title": "Params",
            "type": "object"
          }
        },
        "title": "RunRequest",
        "type": "object"
      },
      "SqlRequest": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Row cap for this query (bounded by ODL_QUERY_MAX_ROWS).",
            "title": "Limit"
          },
          "sql": {
            "description": "A single DuckDB SELECT over the snapshot.",
            "minLength": 1,
            "title": "Sql",
            "type": "string"
          }
        },
        "required": [
          "sql"
        ],
        "title": "SqlRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Fleet health, inventory, utilization, insights and blast radius over a multi-hub OpenShift estate - everything sourced from the OCP API, values scrubbed.",
    "title": "Operations Data Layer",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dashboards/{dashboard_id}": {
      "delete": {
        "description": "Remove a saved dashboard.",
        "operationId": "delete_one_api_dashboards__dashboard_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete One",
        "tags": [
          "dashboards"
        ]
      },
      "get": {
        "description": "The full definition: variables, panels, SQL and all.",
        "operationId": "get_one_api_dashboards__dashboard_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get One",
        "tags": [
          "dashboards"
        ]
      },
      "put": {
        "description": "Create or replace a saved dashboard. The id in the path is the identity.",
        "operationId": "put_one_api_dashboards__dashboard_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put One",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/api/dashboards/{dashboard_id}/run": {
      "post": {
        "description": "Run the whole dashboard against one snapshot build.\n\nThe variables' option queries run in the same batch as the panels, so the\nselector and the numbers beside it describe the same moment. A variable\nwith no value yet is not an error - its options come back so the UI can\ndraw the selector, and the panels that need it say they are waiting for it.",
        "operationId": "post_run_api_dashboards__dashboard_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "dashboard_id",
            "required": true,
            "schema": {
              "title": "Dashboard Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Run",
        "tags": [
          "dashboards"
        ]
      }
    },
...
    "/api/query/ask": {
      "post": {
        "description": "Answer a question in SQL: the query, the rows, and how it was read.",
        "operationId": "post_ask_api_query_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Ask",
        "tags": [
          "query"
        ]
      }
    },
    "/api/query/batch": {
      "post": {
        "description": "Run several SELECTs, with variables, against one snapshot build.\n\nThis is what a dashboard is made of, and the two properties that make it\nmore than a loop over `/sql` are worth being explicit about:\n\n  * **one moment.** Every query runs on a cursor of the same connection,\n    so the panels of a page cannot straddle a rebuild and disagree with\n    each other. The `generation` says which build answered.\n  * **one bad query is not a bad request.** A query the guard refuses or\n    DuckDB cannot run comes back as `{\"error\": ..., \"sql\": ...}` under its\n    own id and the others still answer.\n\nVariables are substituted first, as SQL literals, so the guard sees exactly\nthe text that will run - a placeholder with no value is a 400 naming it,\nrather than an empty string quietly changing what the query means.",
        "operationId": "post_batch_api_query_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Batch",
        "tags": [
          "query"
        ]
      }
    },
...
    "/api/query/sql": {
      "post": {
        "description": "Run one read-only SELECT. The response carries the SQL that ran.",
        "operationId": "post_sql_api_query_sql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SqlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Sql",
        "tags": [
          "query"
        ]
      }
    },
...
  }
}